requirements.txt
Django==3.1.2
django-bootstrap-breadcrumbs==0.9.2
django-bootstrap4==2.3.1
python-dateutil==2.8.1
